Skill remap, Berserk debug logging, spawn fix, server scripts, and git setup
- Remap all skill indices from non-standard 1-10 to clean 0-9 (Skilltypes.h,
  combat.cpp, Client.cpp, HG.cpp, Game.cpp, Skill.cfg, SKILLCFG.TXT, Item.cfg)
- Add debug logging for Berserk spell flow (OBJECTMAGIC, COMMONTYPE_MAGIC,
  no-attack-area check, PlayerMagicHandler) to diagnose casting failures
- Add position save/load debug logging in HG.cpp
- Fix _iGetSkillNumber loop to start at index 0 instead of 1
- Fix character spawn position (initial-point fallback logging)
- Reduce server startup delay from 10s to 3s
- Fix STOP_ALL_SERVERS.bat to close all console windows
- Add .gitignore, README, START/STOP server scripts, firewall setup
- Remove tracked binaries (exe, Usefull Programs, Run_Me SQL)
- Add Maps/MAPDATA shared map data, client GPU renderer, build libs

+++ Sources/HG/char/Npc.h
@@ -69,7 +69,7 @@ public:
 	char  m_cAction;
 	char  m_cTurn;
 
-	short m_sAppr2;
+	int  m_sAppr2;
 
 	DWORD m_dwTime;
 	DWORD m_dwActionTime;
